ts refactoring  config
