+++ package.json
@@ -15,5 +15,24 @@
     "build": "lerna run build",
     "test": "node node_modules/.bin/ava",
     "lint": "node node_modules/.bin/eslint ."
+  },
+  "devDependencies": {
+    "@ava/babel": "^1.0.1",
+    "@types/node": "^14.14.0",
+    "@types/node-fetch": "^2.5.7",
+    "@types/test-listen": "^1.1.0",
+    "@typescript-eslint/eslint-plugin": "^4.5.0",
+    "@typescript-eslint/parser": "^4.5.0",
+    "ava": "^3.13.0",
+    "eslint": "^7.11.0",
+    "eslint-config-standard": "^16.0.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "lerna": "^3.22.1",
+    "node-fetch": "^2.6.1",
+    "test-listen": "^1.1.0",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   }
 }
