+++ tsconfig.build.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+        "target": "ES2019",
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "incremental": true
+    },
+    "exclude": [
+        "test",
+        "node_modules"
+    ]
+}
